backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
supabase==2.13.0
httpx==0.27.2
python-multipart==0.0.6
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
gunicorn==21.2.0
numpy==1.26.4
reportlab==4.2.5
